use a single application.py file.
