+++ browser/.ebextensions/options.config
@@ -1,5 +1,4 @@
 option_settings:
   - namespace: aws:elasticbeanstalk:container:python
     option_name: WSGIPath
-    value: eb_application.py
-
+    value: application.py
